requirements.txt
fastapi
uvicorn
sqlalchemy[postgresql]
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
scikit-learn
pandas
numpy
joblib
requests
streamlit
python-dotenv
plotly
